+++ src/kernel/mod.rs
@@ -2,15 +2,17 @@ pub mod epanechnikov;
 pub mod normal;
 pub mod uniform;
 
+use crate::internal::Float;
+
 pub trait Kernel {
-    fn pdf(&self, x: f64) -> f64;
+    fn pdf(&self, x: Float) -> Float;
 }
 
 impl<T> Kernel for T
 where
-    T: Fn(f64) -> f64,
+    T: Fn(Float) -> Float,
 {
-    fn pdf(&self, x: f64) -> f64 {
+    fn pdf(&self, x: Float) -> Float {
         self(x)
     }
 }
